1.1

+++ python/my_studies/ex007.py
@@ -1,7 +1,7 @@
 #Desenvolva um programa que leia as duas notas de um aluno e calcule e mostre a sua média
 
-nota1 = int(input('Digite a primeira nota: '))
-nota2 = int(input('Digite a segunda nota: '))
+nota1 = float(input('Digite a primeira nota: '))
+nota2 = float(input('Digite a segunda nota: '))
 
 print(f'Nota 1: {nota1}')
 print(f'Nota 2: {nota2}')
